Added connection management and crash handling
Each node has an inbound and an outbound connection.

Outbound conections are automatically started whenever node A sends
the first message to node B and connected when node A receives an
ACK from node B.

Inbound connections are always established when node A receives a
CONNECT message from node B, to which it replies with an ACK.

Each packet carries a connection ID identifier as well as the
sequence number, so whenever a node receives a packet with an
unexpected connection ID, it sends a RESET message to the sender,
which discards all currently pending packets and attempts to
restablish the connection.