Fixed message ordering problem caused by timeouts
Packets now carry the smallest sequence number that hasn't
been ACKd or aborted by the sender. Upon receiving a packet,
the receiver delivers all packets up to
        min(current_sequence, minimum_sequence),
skipping the holes in between. The holes are caused by packets
that have been dropped due to timeouts.

With this change, there is no need to reset a connection due
to timeouts. Business goes on as usual.

These changes are NOT covered in the doc, since we're not delivering
this change as part of assignment 1 (officially), as it is past
the due date.