This change adds the participant termination protocol and timeouts.
At this point what we are missing is log and recovery.

Luciano, we also need to make sure that the code works with your changes. I mean, that the methods are gona get called and all that. Also, we need to figure out how we are going to build the Facebook node on top of this, since you made it no longer a node. My initial idea was for Facebook node to inherit from it and use StartTwoPahseCommit to implement the methods it had that required ACID. Not so sure how that's gona go now.

Lastly, we need to test the hell out of this thing. But again, we need to add recoverability.