Multi-paxos working
Multiple proposers can now queue and negotiate multiple commands
at the same time and eventually all learners will learn the same
chosen commands in the same sequence.

Each proposer queues all commands issued to it and negotiate one
at a time. Multiple proposers may be negotiating commands for the
same slot number. When the chosen value is the command issued by
one proposer, it moves to the next command and next slot. If the
chosen value doesn't match that proposers', it starts a new round
of Paxos for the same command in the next slot.

I added the "dump_values" command which will display all slots
learned so far and the chosen value for that slot.