Of Paxos and failures
Paxos now makes progress and works properly when less than a
majority of nodes are down. As other nodes recover, they
successfully learn the values that were chosen while they were
down. Currently this requires that the new nodes propose a
command. As an optimization, we should look into learning the
values at startup rather than having the node run a Paxos
round for every already-chosen slot value.

This change fixes the stubs to not crash upon RPC errors and
fixes a connection establishment bug when nodes are down for
long periods of time.